Move campgrounds middleware in middleware.js file
