+++ middleware.js
@@ -1,3 +1,8 @@
+const { campgroundSchema } = require("./schemas");
+
+const Campground = require("./models/campground");
+const ExpressError = require("./utils/ExpressError");
+
 const isLoggedIn =  (req, res, next) => {
     if(!req.isAuthenticated()) {
         req.flash("error", "You must be logged in!");
@@ -15,4 +20,24 @@ const storeReturnTo = (req, res, next) => {
     next();
 }
 
-module.exports = {isLoggedIn, storeReturnTo};
+const validateCampground = (req, res, next) => {
+    const { error } = campgroundSchema.validate(req.body); // destructure error from result object
+    if (error) {
+        const message = error.details.map(el => el.message).join(","); // error.details is array of objects so we map 
+        throw new ExpressError(message, 400);                         // over them and make a string of them
+    } else {
+        next();
+    }
+}
+
+const isAuthor = async (req, res, next) => {
+    const { id } = req.params;
+    const campground = await Campground.findById(id);
+    if (!campground.author.equals(req.user._id)) {
+        req.flash("error", "You don't have permission to do that!");
+        return res.redirect(`/campgrounds/${id}`);
+    }
+    next();
+}
+
+module.exports = {isLoggedIn, storeReturnTo, validateCampground, isAuthor};
