+++ views/campgrounds/show.ejs
@@ -51,4 +51,4 @@
             <% } %>     
         </div>
     </div>
-    <a href="/campgrounds" class="btn btn-secondary">All Campgrounds</a>
+    <a href="/campgrounds" class="btn btn-secondary mb-3">All Campgrounds</a>
